pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screen-feedback-agent"
version = "0.1.0"
description = "Convert screen recordings with voice commentary into structured coding tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Martin Priessner" }
]
keywords = ["video", "analysis", "coding", "ai", "gemini", "whisper"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "google-generativeai>=0.3",
    "faster-whisper>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
sfa = "screen_feedback_agent.cli:main"

[project.urls]
Homepage = "https://github.com/mpriessner/screen-feedback-agent"
Repository = "https://github.com/mpriessner/screen-feedback-agent"

[tool.hatch.build.targets.wheel]
packages = ["src/screen_feedback_agent"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=screen_feedback_agent"
